requirements.txt
pandas
psycopg2-binary
tqdm
sklearn
numpy
paramiko
matplotlib
seaborn
scipy
doepy
SALib
sqlparse
aim